fix: namespace disruption type and dependents
